Update paper.md
Some more things I noticed:

Figure 2: I noticed that these properties are not the one mentioned in the text as being important. Would it make sense to add a sentence in the figure caption what the excess Gibbs energy and the endmember activities are useful for?

Figure 3: Would it be worth mentioning the source of this data? I also don’t see the orange data points. 

Figure 4: I am sure the there’s detailed documentation on the planet builder somewhere, would it make sense to link to that here? From the caption alone, it does not really become clear what the planet builder does exactly.